requirements.txt
mysql-connector-python
tqdm
python-dotenv
